project/requirements.txt
Django>=4.0.0,<5.0.0
djangorestframework>=3.13.0,<4.0.0
psycopg2-binary>=2.9.0,<3.0.0
celery>=5.2.0,<6.0.0
redis>=4.0.0,<5.0.0
pandas>=1.4.0,<2.0.0
numpy==1.23.5
openpyxl>=3.0.0,<4.0.0
python-dateutil>=2.8.0,<3.0.0
dj-database-url>=0.5.0,<1.0.0
gunicorn>=20.1.0,<21.0.0
python-dateutil==2.9.0
django-phonenumber-field==8.0.0
phonenumbers==8.13.48
django-cors-headers==4.4.0